Multiple Endpoints implementation
Add a way to use multiple endpoints with this gem.
This is done by adding a parameter to the constructor, endpoints, which
can be either a list of (api_key, app_key) couples or a dict with other
endpoints than the default one.

Tests have been added, and a few of them have been modified to have a
better control on what's happening. (in order to avoid missing api
calls)