Remove multiple endpoints support
This library exists to interact with the API (mute monitors, create
dashboards, ...), and for most of these actions it doesn't make sense to
send them to multiple accounts.

It makes sense for metrics/events/service_checks, and for this usage
dogstatsd-ruby makes more sense. (as it uses the forwarder included with
the agent which can easily send to multiple accounts)