Do not use getlogin() function to obtain username
`getlogin()` obtains the logged-in username by looking up the current tty in `utmp` and reporting the login name found in that utmp record. If Capistrano is ran from Jenkins in the host it returns the user logged in to `pts/0`, which is not the same user as runs the Capistrano. Also Ruby docs recommends avoid using `getlogin()`:
>getlogin → String
>Returns the short user name of the currently logged in user. Unfortunately, it is often rather easy to fool ::getlogin.
>
>Avoid ::getlogin for security-related purposes.
>
>If ::getlogin fails, try ::getpwuid.
>
>See the unix manpage for getpwuid(3) for more detail.
http://ruby-doc.org/stdlib-2.2.0/libdoc/etc/rdoc/Etc.html#method-c-getlogin