Add additional info on options
Metric options like tags are buried in the code documentation and the
issue in #227 is around seeing this in the README for a good example to
use. These may need to change for future versions of ruby because of
keyword argument changes and thus this code will throw warnings anyway
in 2.7 and above versions of Ruby.